# openapi/polaris-mail.yaml
openapi: 3.1.0
info:
  title: polaris-mail API
  version: 0.2.0
  description: |
    HMAC-authenticated REST surface for polaris-mail. All requests are signed with
    `X-Polaris-Sig: <hex(HMAC-SHA256(secret, canonical))>` where canonical is
    `polaris-api\n<METHOD>\n<path>\n<canonical-query>\n<X-Polaris-Ts>\n<X-Polaris-Nonce>\n<sha256-hex(body)>`.

    Webhooks delivered to subscribers (the WebhookEnvelope schema below) are signed
    `X-Polaris-Sig: <hex>` over the canonical string
    `polaris-webhook\nPOST\n<path>\n<canonical-query>\n<ts>\n<nonce>\n<sha256-hex(body)>`.
    Consumers should verify with the SDK (`@polaris-mail/sdk-node`'s `verifyWebhook`
    or `polariswebhook.VerifyWebhook` from `packages/sdk-go`).
servers:
  - url: https://polaris-mail-api.workers.dev
security:
  - HMACAuth: []
components:
  securitySchemes:
    HMACAuth:
      type: apiKey
      in: header
      name: X-Polaris-Sig
...
  schemas:
    Ulid:
      type: string
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
    KeyScope:
      type: string
      enum:
        - send
        - 'messages:read'
        - 'imap_bridge:read'
        - 'admin:rotate'
        - 'admin:read'
        - 'admin:impersonate'
    Operator:
      type: object
      required: [id, name, email, ssh_pubkey_fp_sha256, api_key_id, role, created_at, updated_at]
      properties:
        id: { $ref: '#/components/schemas/Ulid' }
        name: { type: string, minLength: 1, maxLength: 120 }
        email: { type: string, format: email }
        ssh_pubkey:
          type: string
          description: Full authorized_keys line. Omitted from list responses.
        ssh_pubkey_fp_sha256:
          type: string
          pattern: '^SHA256:[A-Za-z0-9+/]{43}$'
          description: SHA256 fingerprint matching `ssh-keygen -E sha256 -lf`.
        api_key_id: { $ref: '#/components/schemas/Ulid' }
        role: { type: string, enum: [admin, operator, readonly] }
        disabled_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time, nullable: true }
    OperatorCreateRequest:
      type: object
      required: [name, email, ssh_pubkey, ssh_pubkey_fp_sha256]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        email: { type: string, format: email }
        ssh_pubkey: { type: string }
        ssh_pubkey_fp_sha256:
          type: string
          pattern: '^SHA256:[A-Za-z0-9+/]{43}$'
        role: { type: string, enum: [admin, operator, readonly], default: operator }
        scopes:
          type: array
          items: { $ref: '#/components/schemas/KeyScope' }
          minItems: 1
          default: ['admin:read']
        rate_limit_per_min: { type: integer, minimum: 1, maximum: 100000, default: 600 }
    OperatorIssueResponse:
      type: object
      required: [api_key_id, api_key_prefix, api_key_secret, login_token]
      properties:
        operator: { $ref: '#/components/schemas/Operator' }
        api_key_id: { $ref: '#/components/schemas/Ulid' }
        api_key_prefix: { type: string }
        api_key_secret:
          { type: string, description: 'Returned ONCE — store via polaris-mail login.' }
        login_token:
          type: string
          pattern: '^polaris_[0-9A-HJKMNP-TV-Z]{26}\..+$'
          description: 'Convenience bundle (polaris_{key_id}.{secret}) for `polaris-mail login`.'
    OperatorUpdateRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        role: { type: string, enum: [admin, operator, readonly] }
        disabled: { type: boolean }
    OperatorRotatePubkeyRequest:
      type: object
      required: [ssh_pubkey, ssh_pubkey_fp_sha256]
      properties:
        ssh_pubkey: { type: string }
        ssh_pubkey_fp_sha256: { type: string, pattern: '^SHA256:[A-Za-z0-9+/]{43}$' }
...
paths:
...
  /v1/admin/operators:
    get:
      summary: List operators
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Operator' }
    post:
      summary: 'Create operator (atomically: principals + api_keys + operators rows)'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorCreateRequest' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorIssueResponse' }
        '409':
          description: email or fingerprint already in use
  /v1/admin/operators/lookup:
    get:
      summary: Lookup operator by SSH pubkey fingerprint (Wish hot path)
      parameters:
        - name: fingerprint
          in: query
          required: true
          schema: { type: string, pattern: '^SHA256:[A-Za-z0-9+/]{43}$' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  operator: { $ref: '#/components/schemas/Operator' }
        '404': { description: not found }
  /v1/admin/operators/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/Ulid' } }
    get:
      summary: Show one operator (includes full SSH pubkey)
      responses: { '200': { description: ok } }
    patch:
      summary: Update operator name / role / disabled flag
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorUpdateRequest' }
      responses: { '200': { description: ok } }
    delete:
      summary: 'Soft-disable operator (alias for PATCH with disabled=true)'
      responses: { '200': { description: ok } }
  /v1/admin/operators/{id}/rotate-key:
    post:
      summary: Rotate the operator's underlying api_key (returns new bearer once)
      parameters:
        - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/Ulid' } }
      responses:
        '200':
          description: rotated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorIssueResponse' }
  /v1/admin/operators/{id}/rotate-pubkey:
    post:
      summary: Replace the operator's SSH pubkey (api_key unchanged)
      parameters:
        - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/Ulid' } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorRotatePubkeyRequest' }
      responses: { '200': { description: ok } }
...